Allow override of SMTP settings per site
This allows for the SMTP settings to be changed on a per-site basis through the config in the database. For some sites it should be possible to use a e-mail provider, depending on the kind of site.

Currently we are only supporting sending e-mail by SMTP, so if we change this support to include other methods of sending mail, we should also make this configuration more generic.