feat(widget-delivery): add `preview` route
The /widget/preview route allows you to set a `Widget-Config` header containing a JSON representation of the config you want to pass to the widget. A `widgetType` matching a key in the `widgetSettingsMapping` must be provided, otherwise an error will be returned. All keys except for the `widgetType` will be passed on to the config.