fix(widget): JSON is now escaped before passing to client
There were some errors with the raw-resource widget in particular, where using `"` or ``` would lead to JSON parsing errors on the server-side. By escaping at a later point we can use the widgetConfig on the server side as intended, and just before passing it to the client we escape the backslashes and quotes correctly.