feat(cms-server): make widget preview configurable
The widgets now extend a new `base-widget`, which extends the old
`@apostrophecms/widget-type`, but allows us to set the `preview` option
for all widgets through a new ENV var `DISABLE_WIDGET_PREVIEW`. By
default `preview` will be set to `true`, which is the default behavior
for Apostrophe.

Because this can lead to a lot of requests in a small timeframe, we want
 the option to disable the previews, so administrators will not be
 blocked by the WAF.