fix: update `knox` dependency
The old `knox` version was 0.9.2 and was using a vulnerable version
of `debug`. By updating to `0.9.3` -- which is only found on Github,
not on npm -- we no longer use a vulnerable version of `debug`.

I opted to use a vendored version of the `knox` dependency due to it
being hosted on Github, which means it could be removed at any time.